[Host_calls]: reorganized. Added runtime-interface
